remove all the JS magic when setting a nick in saved orders index.
It was causing weird problems because browsers were not picking up
the cookie set by JS, making it only work sometimes. The new method
is much simpler, but also slower most likely. This option shouldn’t
be used that often, so it’s not worth the extra effort to make it
fast.