do not test Ruby 2.0.0 anymore
It generally works with 1.9.3 and 2.1.0, so I’m going to
assume it works with 2.0, too. The tests are very brittle
and often fail for random reasons which are hard to debug.

Since the main aim of the tests is to ensure it works
correctly with the pizza.de site, the removed version
doesn’t cost us much.