search subcategories (fixes #5)
This patch doesn’t have a test because capybara-webkit has weird
event handling, which seems to fire for the previous page load
instead of the current one. This leads to the last page not being
searched. Unfortunately the example shop only has subpages on the
last top category, so it doesn’t fire.

It works fine in Firefox/Chrome, so I’m committing anyway.

Late night commits… yaaaay