preload all sub pages on replay
This change introduces unneeded requests on most replays
since the items might be on the first few pages, and the
latter ones would not be needed in the old scheme.

However, when subpages are not cached by HipsterPizza the
time taken to replay an order is quite high, since there’s
the RTT from browser to HipsterPizza and from HP to pizza.de.
It’s assumed that desktop browsers are powerful enough to
handle the many concurrent XHRs and speed up replay
in most cases. It’s disabled on mobile browsers as a
precaution (i.e. haven’t actually tested).